% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToSeuratMetadata.R
\name{addToSeuratMetadata}
\alias{addToSeuratMetadata}
\title{Title: Export infected cell information into the Seurat Object's meta data}
\usage{
addToSeuratMetadata(reads_table, meta_data, library_type = NULL)
}
\arguments{
\item{x}{A table with hiv+ cells metadata.}
}
\value{
A table containing the seurat object metadata with infected cells information
}
\description{
This function takes a a table with hiv+ cell information: barcodes, number of transcripts/dna and sample
and uses  barcodes and sample information to include hiv+ cells data into the seurat object meta data.
}
\examples{
addToSeuratMetadata()
}
