% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetadata.R
\name{loadMetadata}
\alias{loadMetadata}
\title{Title: Load Metadata}
\usage{
loadMetadata(file)
}
\arguments{
\item{x}{A csv file.}
}
\value{
A data.table with the sample metadata.
}
\description{
This function takes a csv file with metadata information and loads it into the workspace as a data.table.
The csv must have the following columns: library_type, sample, read1, read2, index, whitelist.
}
\examples{
loadMetadata("mymetadata.csv")
}
