% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPairedEnd.R
\name{loadPairedEnd}
\alias{loadPairedEnd}
\title{Title: Load Single End reads}
\usage{
loadPairedEnd(
  metadata,
  library_pe = c("Gene Expression", "Chromatin Accessibility"),
  sample = NULL
)
}
\arguments{
\item{metadata}{A table with sample meta data requiered to run virKra}

\item{sample}{Sample id (same as orig.ident from Seurat)}

\item{library_se}{A string indicating data type (e.g.: "Chromatin Accessibility")}
}
\value{
A data.table with reads information
}
\description{
This function takes a fastq files with  PAIRED-END READS mapping to HXB2 genome with Kraken2 and loads it into the workspace as a dataframe
}
\examples{
loadPairedEnd(metadata, library_pe = "Chromatin Accessibility", sample = "sample_id")
}
