% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSingleEnd.R
\name{loadSingleEnd}
\alias{loadSingleEnd}
\title{Title: Load Single End reads}
\usage{
loadSingleEnd(
  metadata,
  library_se = c("Gene Expression", "Chromatin Accessibility"),
  sample = NULL
)
}
\arguments{
\item{metadata}{A table with sample meta data requiered to run virKra}

\item{library_se}{A string indicating data type (e.g.: "Gene Expression")}

\item{sample}{Sample id (same as orig.ident from Seurat)}
}
\value{
A data.table with reads information
}
\description{
This function takes a fastq files with  SINGLE-END READS mapping to HXB2 genome with Kraken2 and loads it into the workspace as a dataframe
}
\examples{
loadSingleEnd(metadata, library_se = "Gene Expression", sample = "sample_id")
}
