% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverBarcode10x.R
\name{recoverBarcode10x}
\alias{recoverBarcode10x}
\title{Title: Recover cells barcodes from indexes}
\usage{
recoverBarcode10x(
  reads_table,
  metadata = NULL,
  cellrang_output_path = NULL,
  nthreads = FALSE
)
}
\arguments{
\item{reads_table}{A table with hiv reads information}

\item{metadata}{A table with hiv+ reads metadata.}

\item{nthreads}{A boolean (T/F) or an integer with the number of threads to use}
}
\value{
A table containing the input table along with the barcode information
}
\description{
This function takes a a table with hiv+ reads and its fastq index files to search for the 10x barcodes of the reads
}
\examples{
recoverBarcode10x()
}
